fix: move window.open handlers from BrowserWindow to WebContents
